[ios] resolve warnings/errors by Swift 6
This CL fixes warnings/errors generated by Swift 6. Swift 6 provides
the safer ways to handle multi-threading. The new warnings/errors
introduced by Swift 6 are mostly around actors and `Task`s:

- Actor isolated call:
https://docs.swift.org/compiler/documentation/diagnostics/actor-isolated-call/
- Sendable closure captures:
https://docs.swift.org/compiler/documentation/diagnostics/sendable-closure-captures/
- Mutable global variable:
https://docs.swift.org/compiler/documentation/diagnostics/mutable-global-variable/
- Conformance isolation:
https://docs.swift.org/compiler/documentation/diagnostics/conformance-isolation/
- Sending risks data race:
https://docs.swift.org/compiler/documentation/diagnostics/sending-risks-data-race/

Bug: 424159227
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6668776
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1478522}

+++ ios/chrome/browser/snapshots/model/image_file_manager.swift
@@ -17,7 +17,7 @@ let kJPEGImageQuality: CGFloat = 1.0
 // Tasks for handling disk (reading an image, writing an image, deleting images, renaming an image,
 // etc.) are executed on a background thread. Callbacks to use UI APIs should be called on the main
 // thread.
-@objcMembers public class ImageFileManager: NSObject {
+@objcMembers public final class ImageFileManager: NSObject, Sendable {
   // Directory where the images are saved.
   private let storageDirectory: URL
   // Scale for snapshot images. It may be smaller than the screen scale in order
@@ -67,7 +67,9 @@ let kJPEGImageQuality: CGFloat = 1.0
 
   // Reads a color image from disk. Reading data for UIImage is executed on the background thread
   // and `completion` is executed on the main thread.
-  func readImage(snapshotID: SnapshotIDWrapper, completion: @escaping (UIImage?) -> Void) {
+  @MainActor
+  func readImage(snapshotID: SnapshotIDWrapper, completion: @escaping @MainActor (UIImage?) -> Void)
+  {
     guard
       let imagePath = imagePath(
         snapshotID: snapshotID, imageType: ImageType.kImageTypeColor)
@@ -77,11 +79,13 @@ let kJPEGImageQuality: CGFloat = 1.0
     }
 
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       let image = UIImage(contentsOfFile: imagePath.path)
       // Call the callback on the main thread.
       mainTaskGroup.enter()
-      DispatchQueue.main.async { [self, image] in
+      DispatchQueue.main.async { [weak self, image] in
+        guard let self = self else { return }
         completion(image)
         // Do not call `backgroundTaskGroup.leave()` here. It causes a deadlock on the main thread
         // if we call `backgroundTaskGroup.wait()` before reaching here.
@@ -101,7 +105,8 @@ let kJPEGImageQuality: CGFloat = 1.0
     }
 
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       guard let data = image.jpegData(compressionQuality: kJPEGImageQuality) else {
         backgroundTaskGroup.leave()
         return
@@ -132,7 +137,8 @@ let kJPEGImageQuality: CGFloat = 1.0
     }
 
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       do {
         if FileManager.default.fileExists(atPath: imagePath.path) {
           try FileManager.default.removeItem(at: imagePath)
@@ -147,7 +153,8 @@ let kJPEGImageQuality: CGFloat = 1.0
   // Removes all images from disk.
   func removeAllImages() {
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       do {
         // Delete the directory storing all images and create a brand new directory with the same
         // storage path.
@@ -222,7 +229,8 @@ let kJPEGImageQuality: CGFloat = 1.0
   // Moves the image in disk from `oldPath` to `newPath`
   func copyImage(oldPath: URL, newPath: URL) {
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       // Copy a file only it's necessary.
       guard FileManager.default.fileExists(atPath: oldPath.path),
         !FileManager.default.fileExists(atPath: newPath.path)
@@ -256,7 +264,8 @@ let kJPEGImageQuality: CGFloat = 1.0
   // `directory`.
   private func createStorageDirectory(directory: URL, legacyDirectory: URL?) {
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       do {
         try FileManager.default.createDirectory(
           at: directory, withIntermediateDirectories: true)
@@ -307,7 +316,8 @@ let kJPEGImageQuality: CGFloat = 1.0
   // after `kGreySnapshotOptimization` feature is enabled by default.
   private func deleteAllGreyImages(directory: URL) {
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       guard FileManager.default.fileExists(atPath: storageDirectory.path) else {
         backgroundTaskGroup.leave()
         return
@@ -345,7 +355,8 @@ let kJPEGImageQuality: CGFloat = 1.0
     }
 
     backgroundTaskGroup.enter()
-    backgroundTaskQueue.async(group: backgroundTaskGroup) { [self] in
+    backgroundTaskQueue.async(group: backgroundTaskGroup) { [weak self] in
+      guard let self = self else { return }
       do {
         // Rename a file only when it's necessary.
         if FileManager.default.fileExists(atPath: oldImagePath.path)
